ML_ANN/pyproject.toml
[project]
name = "ml-ann"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.11.9"
dependencies = [
    "marimo[recommended]>=0.17.7",
    "matplotlib>=3.10.7",
    "mcp>=1",
    "numpy>=2.2.6",
    "onnxruntime>=1.23.2",
    "opencv-python>=4.12.0.88",
    "opencv-python-headless>=4.12.0.88",
    "pillow>=12.0.0",
    "pydantic>=2",
    "pyzmq>=27.1.0",
    "rembg[cli]>=2.0.67",
    "tensorflow>=2.20.0",
]
